Fixed a startup crash
although File.Create() had returned, windows does not create files immediately so trying to access that newly made file causes an exception.

--------------------------------
The method here returns if any files are missing on startup. A better long term solution might be to catch the errors per file so parts of the application can be loaded even if some of it is missing.